src/dispatcher/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
google-cloud-tasks==2.15.1
